remaining base layouts updated, will do a responsive pass next
